src/requirements.txt
tkinter
requests
beautifulsoup4
mysql-connector-python
netifaces
